// ODAKFin_CApp_Feature/ClientApp/src/app/model/Admin.ts
import { ClassStmt } from "@angular/compiler";

export class MyUsers {

    ID = 0;
    UserID = '';
    UserName = '';
    Password = '';
    DOB = '';
    Gender = 0;
    MartialStatus = 0;
    Address = '';
    /*    BranchID = '';*/
    DivID = 0;
    DepID = 0;
    DesID = 0;
    DOJ = '';
    EmailID = '';
    MobNo = '';
    Name = '';
    Designation = '';
    Department = '';
    Division = '';
    Branch = '';
    City = '';

    UID = 0;
    OfficeLocID = 0;
    OfficeLoc = '';
}
export class CommonValues {

    ID: string;
    Desc: string;

}

export class MyDivision {
    ID: string;
    Name: string;
    CityName: string;
}
export class MyCustomerDropdown {

    ID: 0;
    Name: "";
    CustomerName: "";

}
export class MyCntrTypeDropdown {

    ID: 0;
    Name: "";
    CntrTypes: "";

}
export class MyPortdrodown {
    ID: 0;
    PortName: "";
}
export class MyTerminaldrodown {
    ID: 0;
    SlotTerm: "";
    PortID: "";
}
export class MyGeneraldropdown {
    ID: 0;
    GeneralName: "";
}
export class myTerminal {
    ID: 0;
    PortID: 0;
    TerminalName: "";
}
export class MyAgencyDropdown {

    ID: 0;
    AgencyName: "";


}
export class OfficeLocationGrid {
    UID = 0;
    OfficeLocID = 0;
    OfficeLoc = "";
}